Refactor EntityListWidget to accept api/useLegacy/ontologyId/thingType/parameter props and build apiUrl internally
