+++ packages/react/src/components/widgets/EntityListWidget/EntityListWidget.stories.tsx
@@ -1,11 +1,8 @@
-// EntityListWidget.stories.tsx
-
 import type { Meta, StoryObj } from "@storybook/react";
 import { WrappedEntityListWidget } from "./EntityListWidget";
 import {
   EntityListWidgetStoryArgTypes,
   EntityListWidgetStoryArgs,
-  buildEntityListApiUrl,
 } from "./EntityListWidgetStories";
 
 const meta: Meta = {
@@ -13,12 +10,7 @@ const meta: Meta = {
   component: WrappedEntityListWidget,
   argTypes: EntityListWidgetStoryArgTypes,
   args: EntityListWidgetStoryArgs,
-  render: (args: any) => {
-    const apiUrl = buildEntityListApiUrl(args);
-    return <WrappedEntityListWidget apiUrl={apiUrl} />;
-  },
 };
-
 export default meta;
 
 type Story = StoryObj<typeof meta>;
